A little reorganization
