+++ lusSTR/test_suite.py
@@ -1,3 +1,12 @@
+#!/usr/bin/env python
+#
+# -----------------------------------------------------------------------------
+# Copyright (c) 2020, Battelle National Biodefense Institute.
+#
+# This file is part of lusSTR (http://github.com/bioforensics/lusSTR)
+# and is licensed under the BSD license: see LICENSE.txt.
+# -----------------------------------------------------------------------------
+
 import pytest
 import lusSTR
 
@@ -14,13 +23,13 @@ import lusSTR
     )
 ])
 def test_get_annotation(sequence, repeat_list, output):
-    final_output = lusSTR.get_annotation(sequence, repeat_list)
+    final_output = lusSTR.annot.get_annotation(sequence, repeat_list)
     assert final_output == output
 
 
 def test_split_by_n():
     sequence = 'AGGTAGGTAGGTCGAACGAATTGG'
-    blocks = list(lusSTR.split_by_n(sequence, n=4))
+    blocks = list(lusSTR.annot.split_by_n(sequence, n=4))
     assert blocks == [
         'AGGT', 'AGGT', 'AGGT', 'CGAA', 'CGAA', 'TTGG'
     ]
@@ -32,22 +41,22 @@ def test_split_string():
         'TATCTATCTATCTATCTATCTATCTATCTATCTA'
     )
     repeats = ['TCTA', 'TCTG']
-    final_output = lusSTR.split_string(sequence, 6, repeats)
+    final_output = lusSTR.annot.split_string(sequence, 6, repeats)
     assert final_output == '[TCTA]3 [TCTG]6 [TCTA]3 TA [TCTA]3 TCA [TCTA]2 TCCATA [TCTA]11'
 
 
 def test_extract():
     s = '[ATCT]3 ATGT [ATCT]12'
     repeat = 'ATCT'
-    final_output = lusSTR.extract(s, repeat)
+    final_output = lusSTR.annot.extract(s, repeat)
     assert str(final_output) == "12"
 
 
 def test_split_sequence_into_two_strings():
     sequence = 'TAGATAGATAGATGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGGTGTGTGTGTGTG'
-    reverse_comp_sequence = lusSTR.rev_complement_anno(sequence)
+    reverse_comp_sequence = lusSTR.annot.rev_complement_anno(sequence)
     repeat_for_split = 'CACA'
-    seq1, seq2 = lusSTR.split_sequence_into_two_strings(reverse_comp_sequence,
+    seq1, seq2 = lusSTR.annot.split_sequence_into_two_strings(reverse_comp_sequence,
                                                                 repeat_for_split)
     assert seq1 == 'CACACACACACA'
     assert seq2 == 'CCTATCTATCTATCTATCTATCTATCTATCTATCTATCTATCTATCTATCATCTATCTATCTA'
@@ -55,7 +64,7 @@ def test_split_sequence_into_two_strings():
 
 def test_rev_complement_anno():
     sequence = 'TAGATAGATAGATGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGGTGTGTGTGTGTG'
-    final_output = lusSTR.rev_complement_anno(sequence)
+    final_output = lusSTR.annot.rev_complement_anno(sequence)
     assert final_output == (
         'CACACACACACACCTATCTATCTATCTATCTATCTATCTATCTATCTATCTATCTATCTATCATCTATCTATCTA'
     )
@@ -63,13 +72,13 @@ def test_rev_complement_anno():
 
 def test_rev_comp_uas_output_bracket():
     foward_strand = '[AGGT]3 [CGAA]2 TTGG'
-    rev_comp_bracket = lusSTR.rev_comp_uas_output_bracket(foward_strand, 4)
+    rev_comp_bracket = lusSTR.annot.rev_comp_uas_output_bracket(foward_strand, 4)
     assert rev_comp_bracket == 'CCAA [TTCG]2 [ACCT]3'
 
 
 def test_loci_need_split_anno():
     sequence = 'TCTATCTATCTATCTATCTATCTATCTATATATCTATCTATCTATCTA'
-    assert lusSTR.loci_need_split_anno(sequence, 4) == '[TCTA]7 TATA [TCTA]4'
+    assert lusSTR.annot.loci_need_split_anno(sequence, 4) == '[TCTA]7 TATA [TCTA]4'
 
 
 @pytest.mark.parametrize('sequence, bracket_form', [
@@ -101,7 +110,7 @@ def test_loci_need_split_anno():
 ])
 def test_D21_bracket(sequence, bracket_form):
     repeats = ["TCTA", "TCTG"]
-    final_output = lusSTR.D21_bracket(sequence, 6, repeats)
+    final_output = lusSTR.annot.D21_bracket(sequence, 6, repeats)
     assert final_output == bracket_form
 
 
@@ -111,17 +120,17 @@ def test_D19_annotation():
     )
     repeats = ["TCTA", "TCTG"]
     repeat_for_split = 'CCTT'
-    reverse_comp_sequence = lusSTR.rev_complement_anno(sequence)
-    final_output = lusSTR.D19_annotation(reverse_comp_sequence, repeats, repeat_for_split)
+    reverse_comp_sequence = lusSTR.annot.rev_complement_anno(sequence)
+    final_output = lusSTR.annot.D19_annotation(reverse_comp_sequence, repeats, repeat_for_split)
     assert final_output == 'CT CTCT TTCT TCTT CTCT [CCTT]14 CCTA CCTT TT CCTT'
 
 
 def test_D1_annotation():
     sequence = 'TAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATGTGTATGTG'
-    reverse_comp_sequence = lusSTR.rev_complement_anno(sequence)
+    reverse_comp_sequence = lusSTR.annot.rev_complement_anno(sequence)
     repeats = ["TCTA", "CATA", "TCTG", "CACA", "CCTA"]
     repeat_for_split = 'CACA'
-    final_output = lusSTR.D1_annotation(reverse_comp_sequence, repeats, repeat_for_split)
+    final_output = lusSTR.annot.D1_annotation(reverse_comp_sequence, repeats, repeat_for_split)
     print(final_output)
     assert final_output == 'CA CATA CACA [TCTA]11'
 
@@ -138,7 +147,7 @@ def test_D1_annotation():
 def test_PentaD_annotation(sequence, bracket_form):
     repeats = ["AAAGA"]
     no_of_repeat_bases = 5
-    final_output = lusSTR.PentaD_annotation(sequence, no_of_repeat_bases, repeats)
+    final_output = lusSTR.annot.PentaD_annotation(sequence, no_of_repeat_bases, repeats)
     assert final_output == bracket_form
 
 
@@ -148,8 +157,8 @@ def test_FGA_anno():
         'TCTTTCTTTCTTTCTTTCTTTCTTTCTTTCTTTCTTTCTTTCTTCCTTCCTTCCTTTCTTTCTTTCTCCTTCCTTCCTTCCTTCC'
     )
     repeats = ["AAAG", "GAAA", "GAAG", "ACAG", "AAAA"]
-    reverse_comp_sequence = lusSTR.rev_complement_anno(sequence)
-    final_output = lusSTR.FGA_anno(reverse_comp_sequence, repeats)
+    reverse_comp_sequence = lusSTR.annot.rev_complement_anno(sequence)
+    final_output = lusSTR.annot.FGA_anno(reverse_comp_sequence, repeats)
     assert final_output == '[GGAA]4 GGAG [AAAG]3 [GAAG]3 [AAAG]15 [ACAG]3 [AAAG]9 AA AAAA [GAAA]4'
 
 
@@ -170,7 +179,7 @@ def test_FGA_anno():
     )
 ])
 def test_traditional_str_allele(sequence, n, n_sub_out, allele):
-    assert lusSTR.traditional_str_allele(sequence, n, n_sub_out) == allele
+    assert lusSTR.annot.traditional_str_allele(sequence, n, n_sub_out) == allele
 
 
 @pytest.mark.parametrize(
@@ -189,7 +198,7 @@ def test_traditional_str_allele(sequence, n, n_sub_out, allele):
     ])
 def test_lus_anno(forward_bracket, lus, sec, tert, locus, lus_allele, sec_allele, tert_allele,
                   str_allele):
-    lus_out, sec_out, tert_out = lusSTR.lus_anno(
+    lus_out, sec_out, tert_out = lusSTR.annot.lus_anno(
         forward_bracket, lus, sec, tert, locus, str_allele
     )
     assert str(lus_out) == lus_allele
@@ -200,7 +209,7 @@ def test_lus_anno(forward_bracket, lus, sec, tert, locus, lus_allele, sec_allele
 def test_THO1():
     sequence = 'AATGAATGAATGAATGAATGATGATGAATGAATGAATG'
     repeats = ["AATG"]
-    final_output = lusSTR.TH01_annotation(sequence, repeats)
+    final_output = lusSTR.annot.TH01_annotation(sequence, repeats)
     assert final_output == '[AATG]5 ATG ATG [AATG]3'
 
 
@@ -213,7 +222,7 @@ def test_D21_anno(forward_bracket, lus_allele, sec_allele, tert_allele):
     lus = "TCTA"
     sec = "TCTA"
     tert = "TCTG"
-    final_lus, final_sec, final_tert = lusSTR.lus_D21_anno(forward_bracket, lus, sec, tert)
+    final_lus, final_sec, final_tert = lusSTR.annot.lus_D21_anno(forward_bracket, lus, sec, tert)
     assert str(final_lus) == lus_allele
     assert str(final_sec) == sec_allele
     assert str(final_tert) == tert_allele
@@ -223,6 +232,6 @@ def test_D21_lus_sec():
     sequence = '[TCTA]4 [TCTG]6 [TCTA]3 TA [TCTA]3 TCA [TCTA]2 TCCATA [TCTA]10'
     repeat = 'TCTA'
     tert = 'TCTG'
-    lus_out, sec_out = lusSTR.D21_lus_sec(sequence, repeat, tert)
+    lus_out, sec_out = lusSTR.annot.D21_lus_sec(sequence, repeat, tert)
     assert str(lus_out) == '10'
     assert str(sec_out) == '4'
